backend/requirements.txt
amqp==2.2.2
attrs==17.4.0
billiard==3.5.0.3
celery==4.1.0
Django==2.0.1
django-redis==4.8.0
djangorestframework==3.7.7
fancycompleter==0.8
kombu==4.1.0
pdbpp==0.9.2
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.2
Pygments==2.2.0
pytest==3.3.2
pytest-django==3.1.2
pytz==2017.3
redis==2.10.6
six==1.11.0
vine==1.1.4
wmctrl==0.3
